backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
langchain==0.0.281
langchain-openai==0.0.2
openai==1.3.0
faiss-cpu==1.7.4
pypdf==3.15.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
tiktoken==0.5.1
pydantic==2.4.2
tqdm==4.66.1